敌人AI单独封装
