+++ src/utils/index.ts
@@ -32,7 +32,6 @@ export {
   executePendingItemUse,
   enterBattle,
   attack,
-  processEnemyTurn,
   resetBattle,
   toggleFormation,
   autoExecuteBattlePhase,
